+++ compose_rl/data/__init__.py
@@ -6,6 +6,12 @@ from compose_rl.data.dataloader import (
     build_pairwise_preference_dataloader,
     build_prompt_dataloader,
 )
+from compose_rl.data.math_utils import (
+    is_equiv,
+    last_boxed_only_string,
+    normalize_final_answer,
+    remove_boxed,
+)
 from compose_rl.data.preference_data import (
     finegrained_preference_dataset_collate_fn,
     pairwise_preference_dataset_collate_fn,
@@ -14,12 +20,8 @@ from compose_rl.data.prompt_data import prompt_dataset_collate_fn
 from compose_rl.data.rlvr_utils import (
     extract_gsm8k_answer,
     extract_math_answer,
-    is_equiv,
-    last_boxed_only_string,
-    normalize_final_answer,
     prepare_gsm8k_prompt,
     prepare_math_prompt,
-    remove_boxed,
 )
 
 __all__ = [
